Remove interface file and move interface to the component file
